requirements.txt
streamlit==1.26.1
python-dotenv==0.19.0
langchain==0.1.0
langchain-community==0.0.10
google-generativeai>=0.3.2
langchain-google-genai>=0.0.5
proto-plus==1.22.3